Fixed typo in namespace and removed unused file
