+++ src/Packer.php
@@ -22,10 +22,10 @@ use Hostnet\Component\Resolver\EventListener\CleanCssListener;
 use Hostnet\Component\Resolver\EventListener\UglifyJsListener;
 use Hostnet\Component\Resolver\FileSystem\FileReader;
 use Hostnet\Component\Resolver\FileSystem\FileWriter;
-use Hostnet\Component\Resolver\Import\BuildIn\AngularImportCollector;
-use Hostnet\Component\Resolver\Import\BuildIn\JsImportCollector;
-use Hostnet\Component\Resolver\Import\BuildIn\LessImportCollector;
-use Hostnet\Component\Resolver\Import\BuildIn\TsImportCollector;
+use Hostnet\Component\Resolver\Import\BuiltIn\AngularImportCollector;
+use Hostnet\Component\Resolver\Import\BuiltIn\JsImportCollector;
+use Hostnet\Component\Resolver\Import\BuiltIn\LessImportCollector;
+use Hostnet\Component\Resolver\Import\BuiltIn\TsImportCollector;
 use Hostnet\Component\Resolver\Import\ImportFinder;
 use Hostnet\Component\Resolver\Import\Nodejs\Executable;
 use Hostnet\Component\Resolver\Import\Nodejs\FileResolver;
