+++ test/Import/BuiltIn/LessImportCollectorTest.php
@@ -3,7 +3,7 @@
  * @copyright 2017 Hostnet B.V.
  */
 declare(strict_types=1);
-namespace Hostnet\Component\Resolver\Import\BuildIn;
+namespace Hostnet\Component\Resolver\Import\BuiltIn;
 
 use Hostnet\Component\Resolver\File;
 use Hostnet\Component\Resolver\Import\Import;
@@ -11,7 +11,7 @@ use Hostnet\Component\Resolver\Import\ImportCollection;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \Hostnet\Component\Resolver\Import\BuildIn\LessImportCollector
+ * @covers \Hostnet\Component\Resolver\Import\BuiltIn\LessImportCollector
  */
 class LessImportCollectorTest extends TestCase
 {
