--- test/fixtures/resolve.config.json
@@ -1,20 +1,0 @@
-{
-  "files": [
-    "resolver/ts/import-syntax/main.ts"
-  ],
-  "assets": [
-    "resolver/less/import-syntax/main.less"
-  ],
-  "web-root": "web",
-  "output-folder": "dist",
-  "output-folder-dev": "dev",
-  "node": {
-    "bin" : "/home/ydelange/node-v8.5.0-linux-x64/bin/node",
-    "node_modules" : "../../node_modules"
-  },
-  "loaders": [
-    "less",
-    "ts",
-    "angular"
-  ]
-}
